+++ src/main/java/nl/mvdr/adventofcode/adventofcode2017/day18/ModuloInstruction.java
@@ -23,7 +23,7 @@ class ModuloInstruction extends UpdateRegisterInstruction {
     }
     
     @Override
-    int computeNewValue(int oldValue, State state) {
+    long computeNewValue(long oldValue, State state) {
         return oldValue % getValue(state);
     }
