+++ src/main/java/nl/mvdr/adventofcode/adventofcode2017/day18/MultiplyInstruction.java
@@ -22,7 +22,7 @@ class MultiplyInstruction extends UpdateRegisterInstruction {
     }
     
     @Override
-    int computeNewValue(int oldValue, State state) {
+    long computeNewValue(long oldValue, State state) {
         return oldValue * getValue(state);
     }
